+++ src/application/web.py
@@ -6,7 +6,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.openapi.docs import get_swagger_ui_html
 
-from src.infrastructure.di import inject
+from src.infrastructure.di.inject import inject
 from src.infrastructure.utils.config_reader import ConfigReader