+++ src/domain/models/__init__.py
@@ -1,4 +1,0 @@
-from src.domain.models.account.account import Account
-from src.domain.models.profile.profile import Profile
-
-__all__ = ["Account", "Profile"]