+++ src/infrastructure/context/log_management/ilog_manager.py
@@ -1,8 +1,0 @@
-
-from abc import ABC, abstractmethod
-
-class ILogManager(ABC):
-
-    @abstractmethod
-    def set_log(self, message: str, from_file: str, line: int, type: str) -> bool:
-        pass